refactor to have subpackage for richcontext namespace
