+++ .gitignore
@@ -1,9 +1,13 @@
-.ipynb_checkpoints/*
-.DS_Store
-**/.ipynb_checkpoints
-__pycache__
-**~
-
 rc.cfg
 chromedriver
 chromedriver.exe
+
+.ipynb_checkpoints
+.DS_Store
+**~
+
+dist
+build
+__pycache__
+*.pyc
+*.egg-info
