+++ richcontext/scholapi/__init__.py
@@ -1,0 +1,1 @@
+from .scholapi import unpaywall_publication_lookup
